Merge Arctic-Fox commits

+++ netwerk/sctp/src/user_route.h
@@ -10,7 +10,7 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 4. Neither the name of the University nor the names of its contributors
+ * 3. Neither the name of the University nor the names of its contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
  *
@@ -54,7 +54,7 @@ struct sctp_route {
  * retransmission behavior and are included in the routing structure.
  */
 struct sctp_rt_metrics_lite {
-	u_long	rmx_mtu;	/* MTU for this path */
+	uint32_t rmx_mtu;	/* MTU for this path */
 #if 0
 	u_long	rmx_expire;	/* lifetime for route, e.g. redirect */
 	u_long	rmx_pksent;	/* packets sent using this route */
